requirements.txt
boto3
inquirer
pyinstaller
python-dotenv